+++ app/repositories/concentration_repository.py
@@ -1,5 +1,5 @@
 from app.repositories.base_repository import BaseRepository
-from app.models.models import Concentration
+from app.models import Concentration
 
 class ConcentrationRepository(BaseRepository[Concentration]):
     pass
